% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{alpha.estimate}
\alias{alpha.estimate}
\title{process to likelihood}
\usage{
alpha.estimate(tableij, s_h, max.fu)
}
\arguments{
\item{tableij}{tableij}

\item{s_h}{time}

\item{max.fu}{max.fu}
}
\value{
parameters
}
\description{
process to likelihood
}
