% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{My.counts.pb}
\alias{My.counts.pb}
\title{calculate the counts of concordance for variance estimation: using perturbation}
\usage{
My.counts.pb(ss, data.risk)
}
\arguments{
\item{ss}{time}

\item{data.risk}{data}
}
\value{
list of two counts
}
\description{
calculate the counts of concordance for variance estimation: using perturbation
}
