% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{My.counts}
\alias{My.counts}
\title{calculate the counts of concordance for estimation}
\usage{
My.counts(ss, data.risk)
}
\arguments{
\item{ss}{time}

\item{data.risk}{data}
}
\value{
list of two counts
}
\description{
calculate the counts of concordance for estimation
}
