% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{My.loglikehood}
\alias{My.loglikehood}
\title{pseudo partial-likelihood}
\usage{
My.loglikehood(beta, score, totalnum, eventsnum)
}
\arguments{
\item{beta}{beta}

\item{score}{score}

\item{totalnum}{totalnum}

\item{eventsnum}{eventsnum}
}
\value{
negative log likelihood
}
\description{
pseudo partial-likelihood
}
