% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{my.reReg.pb}
\alias{my.reReg.pb}
\title{For variance estimation: using perturbation}
\usage{
my.reReg.pb(w, data, beta1, X)
}
\arguments{
\item{w}{weight}

\item{data}{data set}

\item{beta1}{coefficient}

\item{X}{}
}
\value{
par2.pb A list with optimization results
}
\description{
For variance estimation: using perturbation
}
