% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{my.reReg}
\alias{my.reReg}
\title{Recurrent event regression (proportional means model)}
\usage{
my.reReg(data, beta1, X)
}
\arguments{
\item{data}{data set}

\item{beta1}{coefficient}

\item{X}{}
}
\value{
par2 A list with optimization results
}
\description{
Recurrent event regression (proportional means model)
}
