% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{objective_ee}
\alias{objective_ee}
\title{Objective functions: use estimating equation}
\usage{
objective_ee(beta, input_list)
}
\arguments{
\item{beta}{coefficient}

\item{input_list}{list of inputs}
}
\value{
estimating equation
}
\description{
Objective functions: use estimating equation
}
