% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{objective_ee_pb}
\alias{objective_ee_pb}
\title{Objective functions: use estimating equation with pb}
\usage{
objective_ee_pb(beta, input_list)
}
\arguments{
\item{beta}{coefficient}

\item{input_list}{list of inputs}
}
\value{
estimating equation
}
\description{
Objective functions: use estimating equation with pb
}
