% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{objective_pl}
\alias{objective_pl}
\title{Objective functions: use mimic log partial likelihood}
\usage{
objective_pl(beta, input_list)
}
\arguments{
\item{beta}{coefficient}

\item{input_list}{list of inputs}
}
\value{
negative loglikelihood
}
\description{
Objective functions: use mimic log partial likelihood
}
