% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{objective_pl_pb}
\alias{objective_pl_pb}
\title{Objective functions: use mimic log partial likelihood pb}
\usage{
objective_pl_pb(beta, input_list)
}
\arguments{
\item{beta}{coefficient}

\item{input_list}{list of inputs}
}
\value{
negative log likelihood
}
\description{
Objective functions: use mimic log partial likelihood pb
}
