% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentCIndex.R
\name{TimeDependentCIndex}
\alias{TimeDependentCIndex}
\title{Time-dependent C-index for recurrent events}
\usage{
TimeDependentCIndex(train.data, validate.data, n.var, B = 200)
}
\arguments{
\item{train.data}{training data set}

\item{n.var}{Number of predictors.}

\item{B}{Number of perturbations}

\item{validate.dataInput}{validation data set}
}
\value{
A list with the elements
}
\description{
Time-dependent C-index for recurrent events
}
